WIP #1186 Moved backend whitelisting to execution
